moved testTileScript to PlayMode Tests
