--- Crypto Wars/Assets/Scripts/Test_PlayMode/NewTestScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 40a254b56a348554d9e1d9d671ad29eb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
