+++ Crypto Wars/Assets/Scripts/Test_PlayMode/Tests 1.asmdef
@@ -2,10 +2,7 @@
     "name": "Tests1",
     "rootNamespace": "",
     "references": [
-        "UnityEngine.TestRunner",
-        "UnityEditor.TestRunner",
-        "Unity.TextMeshPro",
-        "Unity.VisualScripting.Core"
+        "Scripts"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
