+++ Crypto Wars/Assets/Scripts/TileScript.cs
@@ -46,7 +46,18 @@ public class Tile : MonoBehaviour
     public void SetMaterial(Material material) {
         rendererReference.material = material;
     }
-    
+
+    public Material GetMaterial()
+    {
+        return rendererReference.material;
+    }
+
+    public void SetTilePosition(int x, int y)
+    {
+        this.BoardXPos = x;
+        this.BoardYPos = y;
+    }
+
     public Vector2 GetTilePosition()
     {
         return new Vector2(BoardXPos, BoardYPos);
